Integrate expandConditionals into expandSums
With the ill-typed outer join operators gone in 3.1, emulateOuterJoins
does not have to interact with Option expansion anymore, so there is no
need to have an `expandConditionals` phase separate from `expandSums`.
Integrating it avoids the use of phase state in `expandSums`.
